Uploaded new server code using nginx, node.js and ssl domains
This update uses node.js to host an API on it’s own subdomain. Static
files and the administration panel is also on it’s own subdomain (using
SSL/HTTPS depending on the situation, https files are not mixed with
http files and the other way around). The master domain is http only
for it not to slow down the browsing experience. Login is ajax and
cookies are stored in the browser using HTML5 web storage, this is so
cookies are not included in all header requests, which should at least
stop the information from being thrown around and should save
bandwidth. This is at the disadvantage of not being able to use the
“secureOnly” parameter of cookies, however, look into cross site
scripting (XSS) attacks. The admin panel is also all ajax, loading
websites as JSON and converting them to HTML. The conversion and editor
in the admin panel is own library specially created for this project.